% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ap}
\alias{ap}
\title{NCAA Men's basketball data rankings}
\format{
\subsection{\code{ap}}{

A data frame with 25 rows and 2 columns:
\describe{
\item{team}{The NCAA men's basketball team}
\item{ap_rank}{The AP rank following the 2023-24 regular season}
}
}
}
\usage{
ap
}
\description{
Collected and cleaned AP rankings for NCAA men's basketball following the 2023-24
regular season basketball data.
}
\keyword{datasets}
