% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_ap_rankings}
\alias{clean_ap_rankings}
\title{Clean NCAA men's basketball AP rankings}
\usage{
clean_ap_rankings(ap_data)
}
\arguments{
\item{ap_data}{tibble returned by \code{\link[=read_ap_rankings]{read_ap_rankings()}}}
}
\value{
A tibble
}
\description{
Cleans data as read in by \code{\link[=read_ap_rankings]{read_ap_rankings()}}. Using this function directly is
discouraged. Instead, load the ap rankings directly with \code{ap}. For more information,
run \code{?ap}.
}
