% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{clean_ncaa_data}
\alias{clean_ncaa_data}
\title{Clean NCAA data}
\usage{
clean_ncaa_data(ncaa_data)
}
\arguments{
\item{ncaa_data}{A tibble returned from \code{\link[=read_ncaa_data]{read_ncaa_data()}}}
}
\value{
A tibble with team, date, and game result columns
}
\description{
Cleans NCAA data that was read using \code{\link[=read_ncaa_data]{read_ncaa_data()}}
}
