% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{game_data_summary}
\alias{game_data_summary}
\title{NCAA game data summary}
\usage{
game_data_summary(ncaa_data)
}
\arguments{
\item{ncaa_data}{NCAA data loaded and cleaned}
}
\value{
A tibble with four columns: team, w, l, wr
}
\description{
Returns a summary of wins, losses, and win-ratio given ncaa data. Use \code{ncaam} or
\code{pac12} as data sources.
}
