% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pac12}
\alias{pac12}
\title{PAC-12 Men's basketball data}
\format{
\subsection{\code{pac12}}{

A data frame with 240 rows and 7 columns:
\describe{
\item{team}{The team representing this row of data}
\item{date}{The game date}
\item{home_away}{Whether the \code{team} is home, away, or something else. NA represents
something else}
\item{opp}{The \code{team}'s oppponent in this game}
\item{result}{The game result from the persepctive of the \code{team}}
\item{team_score}{The \code{team}'s score}
\item{opp_score}{The \code{opp}'s score}
}
}
}
\usage{
pac12
}
\description{
Collected and cleaned NCAA men's 2023-24 regular season basketball data for PAC-12
intra-conference play only. This data is ready to be prepared for ranking using a
\verb{ranking_table_prep_*()} function. The data is at the team-game level. That is, there
are two rows per game: one per team.
}
\keyword{datasets}
