% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{plot_lambdas}
\alias{plot_lambdas}
\title{Plot lambdas from iteration}
\usage{
plot_lambdas(eigenvalues)
}
\arguments{
\item{eigenvalues}{The result of \code{\link[=eigenvalues_power_method]{eigenvalues_power_method()}}}
}
\value{
A ggplot2 plot
}
\description{
We can use this function to view the results of teh power method. When we view this
plot we should see our lambda converging.
}
