% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{plot_rank_comparison}
\alias{plot_rank_comparison}
\title{Plot rankings compared against AP rankings}
\usage{
plot_rank_comparison(comp_table)
}
\arguments{
\item{comp_table}{A tibble containing at least two columns: \code{rank} and \code{ap_rank}}
}
\value{
A ggplot2 plot
}
\description{
We can use this function to view the results of the power method. When we view this
plot we may notice similarity between the rankings computed and the AP ranking
}
