% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_table_prep.R
\name{ranking_table_prep_direct_win_loss}
\alias{ranking_table_prep_direct_win_loss}
\title{Ranking table preparation}
\usage{
ranking_table_prep_direct_win_loss(ncaa_data)
}
\arguments{
\item{ncaa_data}{Cleaned NCAA data returned from \code{\link[=clean_ncaa_data]{clean_ncaa_data()}}}
}
\value{
A ranking table ready to be passed to \code{\link[=to_matrix]{to_matrix()}}
}
\description{
Given cleaned NCAA data, compute a ranking table using various methods
}
\details{
\code{ranking_table_direct_win_loss()} assigns a 1 for each victory, a 0.5 for
each tie, and a 0 for loss, then adds the results
}
