% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{rankings_from_eigenvalues}
\alias{rankings_from_eigenvalues}
\title{Compute rankings from eigenvalues}
\usage{
rankings_from_eigenvalues(eigenvalues, name_col = "team")
}
\arguments{
\item{eigenvalues}{The result of \code{\link[=eigenvalues_power_method]{eigenvalues_power_method()}}}

\item{name_col}{Column to store names of \code{evs}, if applicable}
}
\value{
A tibble containing the eigenvalues, rank, and a name column, if \code{evs} is
named
}
\description{
Takes a vector of eigenvalues and returns data.frame of eigenvalues and corresponding
ranks
}
