% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_ap_rankings}
\alias{read_ap_rankings}
\title{Read NCAA men's basketball rankings}
\usage{
read_ap_rankings()
}
\value{
A tibble with one column
}
\description{
Returns unprocessed AP rankings of NCAA men's basketball temas following the 2023-24
season. Using this function directly is discouraged. Instead, load the ap rankings
directly with \code{ap}. For more information, run \code{?ap}.
}
