% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{read_ncaa_data}
\alias{read_ncaa_data}
\title{Read NCAA basketball data}
\usage{
read_ncaa_data(league)
}
\arguments{
\item{league}{The league ID corresponding to the sheet name. Currently, only "NCAAM"
is supported.}
}
\value{
A tibble
}
\description{
Reads googlesheet containing collected data from
\href{https://stathead.com/basketball/cbb/team-game-finder.cgi?request=1&comp_type=reg&game_status=1&order_by=date&match=team_game&year_max=2024&order_by_asc=1&timeframe=seasons&comp_id=NCAAM&year_min=2024}{sports-reference.com}.
}
\details{
Calling this function directly will cause you to authenticate using your google sheets
account. You may also access the raw data at \code{ncaam_raw}
}
