% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert to matrix}
\usage{
to_matrix(tbl)
}
\arguments{
\item{tbl}{a table resulting from one of the \verb{ranking_table_*()} family of functions}
}
\value{
A square matrix with named rows
}
\description{
Convert to matrix
}
