% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{view_subset}
\alias{view_subset}
\title{Subset a data.frame by columns and rows}
\usage{
view_subset(data, column_subset, id_col = "team")
}
\arguments{
\item{data}{A data.frame}

\item{column_subset}{The columns + rows to be chosen}

\item{id_col}{The name of the column containing the values that line up with column
names}
}
\value{
A data.frame
}
\description{
This function takes a data.fram containing a column that lines up to column names, and
filters to where the column are in the selected values, and only shows the column
names matching the selected values.
}
